+++ plano/src/com/hendraanggrian/plano/TrimSize.kt
@@ -3,8 +3,8 @@ package com.hendraanggrian.plano
 import java.io.Serializable
 
 class TrimSize(
-    val x: Double,
-    val y: Double,
-    override val width: Double,
-    override val height: Double
+    val x: Float,
+    val y: Float,
+    override val width: Float,
+    override val height: Float
 ) : Size, Serializable
